Get inputs from MRIcron/npm file

Function to load data from a previous analysis in MRIcron/npm in a ready format for use in lesymap

lesyload_mricron(valfile, imageFolder = NA, returnFilenames = F,
  checkHeaders = T, showInfo = T)

Arguments

valfile
mricron filename with extention *.val. The function will search for images in the same folder where valfile is located, unless you specify imageFolder. If any of the files listed in the .val file are not found in the folder, an error will be displayed.

imageFolder
(default=NA) folder to look for the image files

returnFilenames
(default=FALSE) By default the function will load the images in memory to speed up things in lesymap. This may require too much RAM memory in some cases, and you may want to use filenames instead, which requires less memory but is slower in lesymap.

checkHeaders
(default=TRUE) Headers will be checked to make sure all images have the same dimension/origin/resolution, etc.

showInfo
(default=TRUE) show information upon successful load

Value

List with the following information lesions - list of antsImages or vector of filenames behavior - vector of behavioral scores