Brunner-Munzel tests for symptom mapping (fast)

Lesion to symptom mapping performed on a prepared matrix. Brunner-Munzel tests are performed using each column of the matrix to split the behavioral scores in two groups. This function relies on a compiled version for fast processing.

lsm_BMfast(lesmat, behavior, permuteNthreshold = 9,
  alternative = "greater", statOnly = FALSE, nperm = 1000,
  npermBM = 20000, FWERperm = FALSE, v = 1, pThreshold = 0.05,
  permuteAllVoxelsBM = FALSE, showInfo = FALSE, ...)

Arguments

lesmat
binary matrix (0/1) of voxels (columns) and subjects (rows).

behavior
vector of behavioral scores.

permuteNthreshold
(default=9) Voxels lesioned in fewer than 9 subjects may yield incorrect p-values with Brunner-Munzel tests, so they need to identify pvalues through individualized permutations. This parameter sets the threshold to find which voxels need permutations. See Medina et al (2010).

alternative
(default="greater") It is assumed that healthy voxels (0) have greater behavioral scores. If your data follow an inverted relationship choose "less" or "two.sided".

statOnly
logical (default=FALSE), skips some computations, mostly for internal use to speed up some things.

nperm
(default=1000) Number of permutations to perform on entire volumes when needed for multiple comparisons corrections (i.e., in FWERperm).

npermBM
(default=20000) Number of permutations to perform at every single voxel below permuteNthrehsold. Note, this argument is different from nperm, which controls volume-based permutations to perform multiple comparison corrections with FWERperm.

FWERperm
logical (default=FALSE) whether to perform permutation based FWER thresholding.

v
(default=1) which voxel to record at each permutation with FWERperm. All software use the peak voxel (v=1), but you can choose a voxel further down the list to relax the threshold (i.e., v=10 for 10 highest voxel) (see Mirman (2017)).

pThreshold
(default=0.05) what threshold to use for FWER

permuteAllVoxelsBM
(default=FALSE) whether to force the permutation-based p-value calulation for all voxels, instead of applying only to voxels below permuteNthrehsold. Setting this option to TRUE will force all voxels undergo permutation-based p-value calculation.

showInfo
display info messagges when running the function.

...
other arguments received from lesymap.

Value

List of objects returned:

statistic - vector of statistical values

pvalue - vector of pvalues

zscore - vector of zscores

perm.vector - (optional) vector of permuted statistics

perm.FWERthresh - (optional) permutation threshold established from the distribution of perm.vector

Examples

{
set.seed(123)
lesmat = matrix(rbinom(200,1,0.5), ncol=2)
set.seed(123)
behavior = rnorm(100)
result = lsm_BMfast(lesmat, behavior)
}