Regression tests for symptom mapping (fast)

Lesion to symptom mapping performed on a prepared matrix. Regressions are performed between behavior and each column of the lesmat matrix. Fast function based on compiled code.

lsm_regresfast(lesmat, behavior, covariates = NA, FWERperm = F,
  nperm = 1000, v = 1, pThreshold = 0.05, clusterPerm = F,
  mask = NA, voxindx = NA, samplemask = NA,
  clusterPermThreshold = 0.05, showInfo = T, ...)

Arguments

lesmat
matrix of voxels (columns) and subjects (rows).

behavior
vector of behavioral scores.

covariates
(default=NA) vector of matrix of covariates.

FWERperm
logical (default=FALSE) whether to run permutation based FWER thresholding.

nperm
Number of permutations to perform when needed.

v
(default=1) what voxel to record for FWER thresholding.

pThreshold
(default=0.05) Voxel-wise threshold.

clusterPerm
logical (default=FALSE), whether to perform permutation based cluster thresholding.

mask
(default=NA) antsImage reference mask used for cluster computations.

voxindx
(default=NA) indices of voxels to put in mask

samplemask
(default=NA) antsImage used to extract voxels back in a matrix.

clusterPermThreshold
(default=0.05) threshold for cluster selection after obtaining cluster size distrubution.

showInfo
display info messagges when running the function.

...
other arguments received from lesymap.

Value

List of objects returned:

statistic - vector of statistical values

pvalue - vector of pvalues

zscore - vector of zscores

perm.vector - (optional) vector of permuted statistics

perm.FWERthresh - (optional) permutation threshold established from the distribution of perm.vector

perm.clusterThreshold - (optional) permutation threshold established from the distribution of perm.vector

Examples

{
set.seed(123)
lesmat = matrix(rbinom(200,1,0.5), ncol=2)
set.seed(123)
behavior = rnorm(100)
result = lsm_regresfast(lesmat, behavior)
}