Support Vector Regression for symptom mapping

Lesion to symptom mapping performed on a prepared matrix. The SVR method is used. The function relies on the svm function of the e1071 package. The analysis follows a similar logic found in the SVR-LSM code published by Zhang (2015). After a first run of SVM, p-values are established with a permutation procedures as the number of times weights are randomly exceeded in permutations. The returned p-values are not corrected for multiple comparisons.

lsm_svr(lesmat, behavior, SVR.nperm = 10000,
  SVR.type = "eps-regression", SVR.kernel = "radial", SVR.gamma = 5,
  SVR.cost = 30, SVR.epsilon = 0.1, showInfo = TRUE, ...)

Arguments

lesmat
matrix of voxels (columns) and subjects (rows).

behavior
vector of behavioral scores.

SVR.nperm
(default=10,000) number of permutations to run to estimate p-values. Note, these p-values are uncorrected for multiple comparisons.

SVR.type
(default='eps-regression') type of SVM to run, see svm.

SVR.kernel
(default='radial') type of kernel to use, see svm

SVR.gamma
(default=5) gamma value, see svm.

SVR.cost
(default=30) cost value, see svm.

SVR.epsilon
(default=0.1) epsilon value, see svm.

showInfo
logical (default=TRUE) display messages

...
other arguments received from lesymap.

Value

List of objects returned:

statistic - vector of statistical values

pvalue - vector of pvalues